chore(deps): update nitro-codegen and react-native-nitro-modules to version 0.26.2
This commit updates the dependencies for nitro-codegen and react-native-nitro-modules in the bun.lock, package.json, and example/package.json files. Additionally, it bumps the AppleAuth pod version to 0.2.1 in the Podfile.lock and updates various React Native components to version 0.77.1 in the Podfile.lock. A new git attribute flag is also added in nitro.json to indicate generated files.